perf: use search bar #891
